Rename files to have generator prefix.
