+++ src/generatorprofiletools.cpp
