+++ src/generatorprofiletools.h
